Исправляет ошибки
