+++ js/create-ads.js
@@ -1,5 +1,27 @@
-import {getRandomNumber, getRandomFractional, getRandomArrayElement, getRandomArrayElements} from './util.js';
-import {TYPES, TIMES, FEATURES, PHOTOS, Coordinates, ADS_COUNT} from './data.js';
+import {
+  getRandomNumber,
+  getRandomFractional,
+  getRandomArrayElement,
+  getRandomArrayElements
+} from './util.js';
+
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const TIMES = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+const Coordinates = {
+  LAT_MIN: 35.65000,
+  LAT_MAX: 35.70000,
+  LNG_MIN: 139.70000,
+  LNG_MAX: 139.80000
+};
+
+const ADS_COUNT = 10;
 
 // Функция для создания объекта (объявления)
 
@@ -34,4 +56,4 @@ const createAds = (index) => {
 
 const generateAds = () => Array.from({ length: ADS_COUNT }, (_, index) => createAds(index));
 
-generateAds(ADS_COUNT);
+export {generateAds, ADS_COUNT};
