--- js/data.js
@@ -1,19 +1,0 @@
-const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
-const TIMES = ['12:00', '13:00', '14:00'];
-const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
-];
-
-const Coordinates = {
-  LAT_MIN: 35.65000,
-  LAT_MAX: 35.70000,
-  LNG_MIN: 139.70000,
-  LNG_MAX: 139.80000
-};
-
-const ADS_COUNT = 10;
-
-export {TYPES, TIMES, FEATURES, PHOTOS, Coordinates, ADS_COUNT};
