+++ js/main.js
@@ -1,1 +1,4 @@
-import './create-ads.js';
+import {generateAds, ADS_COUNT} from './create-ads.js';
+
+generateAds(ADS_COUNT);
+
